services/gateway_service/pyproject.toml
[project]
name = "gateway-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "python-multipart",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "common",
]

[tool.uv.sources]
common = { workspace = true }
